+++ packages/paste-core/components/button/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.18](https://github.com/twilio-labs/paste/compare/@twilio-paste/button@1.1.17...@twilio-paste/button@1.1.18) (2020-02-20)
+
+**Note:** Version bump only for package @twilio-paste/button
+
+
+
+
+
 ## [1.1.17](https://github.com/twilio-labs/paste/compare/@twilio-paste/button@1.1.16...@twilio-paste/button@1.1.17) (2020-02-19)
 
 **Note:** Version bump only for package @twilio-paste/button