+++ packages/paste-core/primitives/modal-dialog/CHANGELOG.md
@@ -3,14 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.0.2](https://github.com/twilio-labs/paste/compare/@twilio-paste/primitive-modal-dialog@0.0.1...@twilio-paste/primitive-modal-dialog@0.0.2) (2019-10-29)
-
-**Note:** Version bump only for package @twilio-paste/primitive-modal-dialog
+# 0.1.0 (2020-02-19)
 
 
+### Bug Fixes
+
+* **modal-dialog-primitive:** rename exports ([fcdd574](https://github.com/twilio-labs/paste/commit/fcdd574a868b398abab9b1cfe0606772ae43123b))
 
 
+### Features
+
+* **modal-dialog-primitive:** add package ([ba350fc](https://github.com/twilio-labs/paste/commit/ba350fc1541be23f7aa4920e3f618386693a9e13))
+
+
+
+
+
+## [0.0.2](https://github.com/twilio-labs/paste/compare/@twilio-paste/modal-dialog-primitive@0.0.1...@twilio-paste/modal-dialog-primitive@0.0.2) (2019-10-29)
+
+**Note:** Version bump only for package @twilio-paste/modal-dialog-primitive
 
 ## 0.0.1 (2019-08-15)
 
-**Note:** Version bump only for package @twilio-paste/primitive-modal-dialog
+**Note:** Version bump only for package @twilio-paste/modal-dialog-primitive